@@ -9,35 +9,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.2",
-    "@angular/common": "^17.0.2",
-    "@angular/compiler": "^17.0.2",
-    "@angular/core": "^17.0.2",
-    "@angular/forms": "^17.0.2",
-    "@angular/platform-browser": "^17.0.2",
-    "@angular/platform-browser-dynamic": "^17.0.2",
-    "@angular/router": "^17.0.2",
+    "@angular/animations": "^17.2.1",
+    "@angular/common": "^17.2.1",
+    "@angular/compiler": "^17.2.1",
+    "@angular/core": "^17.2.1",
+    "@angular/forms": "^17.2.1",
+    "@angular/platform-browser": "^17.2.1",
+    "@angular/platform-browser-dynamic": "^17.2.1",
+    "@angular/router": "^17.2.1",
     "@fontsource/roboto": "^5.0.8",
     "@ngneat/effects-ng": "^3.1.3",
-    "@ngneat/elf": "^2.4.0",
+    "@ngneat/elf": "^2.5.1",
     "@ngneat/elf-devtools": "^1.3.0",
-    "@ngneat/elf-persist-state": "^1.2.0",
+    "@ngneat/elf-persist-state": "^1.2.1",
     "@ngneat/elf-requests": "^1.9.2",
-    "ng-zorro-antd": "^17.1.0",
+    "ng-zorro-antd": "^17.2.0",
     "normalize.css": "^8.0.1",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^17.0.0",
-    "@angular-devkit/build-angular": "^17.0.0",
-    "@angular/cli": "~17.0.0",
-    "@angular/compiler-cli": "^17.0.2",
-    "@types/jest": "^29.5.11",
+    "@angular-builders/jest": "^17.0.1",
+    "@angular-devkit/build-angular": "^17.2.0",
+    "@angular/cli": "~17.2.0",
+    "@angular/compiler-cli": "^17.2.1",
+    "@types/jest": "^29.5.12",
     "jest": "^29.7.0",
-    "jest-preset-angular": "^13.1.4",
+    "jest-preset-angular": "^14.0.3",
     "ng-mocks": "^14.12.1",
-    "typescript": "~5.2.2"
+    "typescript": "~5.3.3"
   }
 }
