@@ -1,6 +1,6 @@
 import { ErrorService } from '$service/error.service'
-import { CommonModule } from '@angular/common'
-import { Component } from '@angular/core'
+
+import { Component, inject } from '@angular/core'
 import { RouterOutlet } from '@angular/router'
 import { NzAlertModule } from 'ng-zorro-antd/alert'
 import { NzButtonModule } from 'ng-zorro-antd/button'
@@ -10,14 +10,12 @@ import { HeaderComponent } from './header/header.component'
 @Component({
   selector: 'gcd-root',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, HeaderComponent, GroupTabsComponent, NzAlertModule, NzButtonModule],
+  imports: [RouterOutlet, HeaderComponent, GroupTabsComponent, NzAlertModule, NzButtonModule],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  error = this.errorService.error
-
-  constructor(private errorService: ErrorService) {}
+  error = inject(ErrorService).error
 
   OnClick(): void {
     window.location.reload()
