@@ -1,14 +1,6 @@
 import { Project } from '$groups/model/project'
-import { CommonModule } from '@angular/common'
-import {
-  ChangeDetectionStrategy,
-  Component,
-  EventEmitter,
-  Input,
-  OnChanges,
-  Output,
-  SimpleChanges
-} from '@angular/core'
+
+import { ChangeDetectionStrategy, Component, EventEmitter, Output, computed, effect, input } from '@angular/core'
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop'
 import { FormControl, ReactiveFormsModule } from '@angular/forms'
 import { NzButtonModule } from 'ng-zorro-antd/button'
@@ -23,7 +15,6 @@ import { debounceTime } from 'rxjs'
   selector: 'gcd-project-filter',
   standalone: true,
   imports: [
-    CommonModule,
     NzIconModule,
     NzInputModule,
     NzTagModule,
@@ -36,28 +27,23 @@ import { debounceTime } from 'rxjs'
   styleUrls: ['./project-filter.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProjectFilterComponent implements OnChanges {
-  @Input({ required: true }) projects: Project[] = []
-  @Input({ required: true }) selectedFilterText: string = ''
+export class ProjectFilterComponent {
+  projects = input.required<Project[]>()
+  selectedFilterText = input.required<string>()
 
   @Output() filterTextChanged = new EventEmitter<string>()
 
+  projectCount = computed(() => new Set(this.projects().map(({ id }) => id)).size)
   searchControl = new FormControl('')
 
   constructor() {
+    effect(() => {
+      this.searchControl.setValue(this.selectedFilterText(), { emitEvent: false })
+    })
+
     this.searchControl.valueChanges
       .pipe(takeUntilDestroyed(), debounceTime(100))
       .subscribe((value) => this.filterTextChanged.next(String(value)))
-  }
-
-  ngOnChanges({ selectedFilterText }: SimpleChanges): void {
-    if (selectedFilterText) {
-      this.searchControl.setValue(this.selectedFilterText, { emitEvent: false })
-    }
-  }
-
-  get projectCount(): number {
-    return new Set(this.projects.map(({ id }) => id)).size
   }
 
   resetSearch(): void {
