@@ -3,13 +3,14 @@ import { GroupId } from '$groups/model/group'
 import { ScheduleWithProjectAndPipeline } from '$groups/model/schedule'
 import { ErrorService } from '$service/error.service'
 import { HttpClient } from '@angular/common/http'
-import { Injectable } from '@angular/core'
+import { Injectable, inject } from '@angular/core'
 import { Observable, catchError, identity, retry, throwError } from 'rxjs'
 import { trackRequestsStatus } from '../store/schedule.store'
 
 @Injectable({ providedIn: 'root' })
 export class ScheduleService {
-  constructor(private http: HttpClient, private errorService: ErrorService) {}
+  private http = inject(HttpClient)
+  private errorService = inject(ErrorService)
 
   getSchedules(groupId: GroupId, withLoader: boolean = true): Observable<ScheduleWithProjectAndPipeline[]> {
     const url = `${location.origin}/api/schedules`
