@@ -1,7 +1,9 @@
 package branch
 
 import (
+	"context"
 	"fmt"
+
 	"github.com/larscom/gitlab-ci-dashboard/branch/mock"
 
 	"testing"
@@ -15,7 +17,7 @@ func TestGetBranchesWith1Page(t *testing.T) {
 		client     = NewClient(mock.NewGitlabClientMock(totalPages, nil))
 	)
 
-	branches, _ := client.GetBranches(1)
+	branches, _ := client.GetBranches(1, context.Background())
 
 	assert.Len(t, branches, 2)
 	assert.Equal(t, "branch-1", branches[0].Name)
@@ -28,7 +30,7 @@ func TestGetBranchesWith2Pages(t *testing.T) {
 		client     = NewClient(mock.NewGitlabClientMock(totalPages, nil))
 	)
 
-	branches, _ := client.GetBranches(1)
+	branches, _ := client.GetBranches(1, context.Background())
 
 	assert.Len(t, branches, 4)
 	assert.Equal(t, "branch-1", branches[0].Name)
@@ -40,7 +42,7 @@ func TestGetBranchesWith2Pages(t *testing.T) {
 func TestGetBranchesWithErrorEmptySlice(t *testing.T) {
 	client := NewClient(mock.NewGitlabClientMock(0, fmt.Errorf("ERROR")))
 
-	branches, _ := client.GetBranches(100)
+	branches, _ := client.GetBranches(100, context.Background())
 
 	assert.Len(t, branches, 0)
 }
