@@ -1,9 +1,13 @@
 package mock
 
-import "github.com/larscom/gitlab-ci-dashboard/model"
+import (
+	"context"
+
+	"github.com/larscom/gitlab-ci-dashboard/model"
+)
 
 type ClientMock struct{}
 
-func (c *ClientMock) GetBranches(projectId int) ([]model.Branch, error) {
+func (c *ClientMock) GetBranches(projectId int, ctx context.Context) ([]model.Branch, error) {
 	return []model.Branch{{Name: "branch-1"}}, nil
 }
