@@ -1,12 +1,16 @@
 package mock
 
-import "github.com/larscom/gitlab-ci-dashboard/model"
+import (
+	"context"
+
+	"github.com/larscom/gitlab-ci-dashboard/model"
+)
 
 type BranchServiceMock struct {
 	Error error
 }
 
-func (s *BranchServiceMock) GetBranchesWithLatestPipeline(projectId int) ([]model.BranchWithPipeline, error) {
+func (s *BranchServiceMock) GetBranchesWithLatestPipeline(projectId int, ctx context.Context) ([]model.BranchWithPipeline, error) {
 	if projectId == 1 {
 		return []model.BranchWithPipeline{
 			{
