@@ -1,7 +1,9 @@
 package group
 
 import (
+	"context"
 	"fmt"
+
 	"github.com/larscom/gitlab-ci-dashboard/group/mock"
 
 	"strings"
@@ -36,7 +38,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client     = NewClient(mock.NewGitlabClientMock(totalPages, nil), cfg)
 		)
 
-		groups, _ := client.GetGroups()
+		groups, _ := client.GetGroups(context.Background())
 
 		assert.Len(t, groups, 2)
 		assert.Equal(t, "group-1", groups[0].Name)
@@ -49,7 +51,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client     = NewClient(mock.NewGitlabClientMock(totalPages, nil), cfg)
 		)
 
-		groups, _ := client.GetGroups()
+		groups, _ := client.GetGroups(context.Background())
 
 		assert.Len(t, groups, 4)
 		assert.Equal(t, "group-1", groups[0].Name)
@@ -65,7 +67,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client       = NewClient(mock.NewGitlabClientMock(totalPages, nil), cfg)
 		)
 
-		groups, _ := client.GetGroups()
+		groups, _ := client.GetGroups(context.Background())
 
 		assert.Len(t, groups, 2)
 		assert.Equal(t, "group-20", groups[0].Name)
@@ -79,7 +81,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client     = NewClient(mock.NewGitlabClientMock(totalPages, nil), cfg)
 		)
 
-		groups, _ := client.GetGroups()
+		groups, _ := client.GetGroups(context.Background())
 
 		assert.Len(t, groups, 2)
 		assert.Equal(t, "group-10", groups[0].Name)
@@ -91,7 +93,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client = NewClient(mock.NewGitlabClientMock(0, fmt.Errorf("ERROR")), cfg)
 		)
 
-		groups, _ := client.GetGroups()
+		groups, _ := client.GetGroups(context.Background())
 
 		assert.Len(t, groups, 0)
 	})
@@ -102,7 +104,7 @@ func TestGroupClientWithConfig(t *testing.T) {
 			client     = NewClient(mock.NewGitlabClientMock(totalPages, nil), cfg)
 		)
 
-		groups, _ := client.GetGroupsById([]int{1, 2})
+		groups, _ := client.GetGroupsById([]int{1, 2}, context.Background())
 
 		assert.Len(t, groups, 2)
 
