@@ -1,13 +1,17 @@
 package mock
 
-import "github.com/larscom/gitlab-ci-dashboard/model"
+import (
+	"context"
+
+	"github.com/larscom/gitlab-ci-dashboard/model"
+)
 
 type ClientMock struct{}
 
-func (c *ClientMock) GetGroupsById(ids []int) ([]model.Group, error) {
+func (c *ClientMock) GetGroupsById(ids []int, ctx context.Context) ([]model.Group, error) {
 	return []model.Group{{Name: "Z"}, {Name: "X"}, {Name: "Y"}}, nil
 }
 
-func (c *ClientMock) GetGroups() ([]model.Group, error) {
+func (c *ClientMock) GetGroups(ctx context.Context) ([]model.Group, error) {
 	return []model.Group{{Name: "C"}, {Name: "A"}, {Name: "B"}}, nil
 }
