@@ -1,6 +1,10 @@
 package mock
 
-import "github.com/larscom/gitlab-ci-dashboard/model"
+import (
+	"context"
+
+	"github.com/larscom/gitlab-ci-dashboard/model"
+)
 
 type ClientMock struct{}
 
@@ -12,7 +16,7 @@ func (c *ClientMock) GetLatestPipelineBySource(projectId int, ref string, source
 	return &model.Pipeline{ProjectId: projectId, Ref: ref, Source: source, Status: "success", Id: 1337}, nil
 }
 
-func (c *ClientMock) GetPipelines(projectId int) ([]model.Pipeline, error) {
+func (c *ClientMock) GetPipelines(projectId int, ctx context.Context) ([]model.Pipeline, error) {
 	return []model.Pipeline{
 		{ProjectId: projectId, Status: "success", Id: 1},
 		{ProjectId: projectId, Status: "failed", Id: 2},
