@@ -1,7 +1,9 @@
 package project
 
 import (
+	"context"
 	"fmt"
+
 	"github.com/larscom/gitlab-ci-dashboard/project/mock"
 
 	"testing"
@@ -15,7 +17,7 @@ func TestGetProjectsWith1Page(t *testing.T) {
 		client     = NewClient(mock.NewGitlabClientMock(totalPages, nil))
 	)
 
-	projects, _ := client.GetProjects(1)
+	projects, _ := client.GetProjects(1, context.Background())
 
 	assert.Len(t, projects, 2)
 	assert.Equal(t, "project-1", projects[0].Name)
@@ -28,7 +30,7 @@ func TestGetProjectsWith2Pages(t *testing.T) {
 		client     = NewClient(mock.NewGitlabClientMock(totalPages, nil))
 	)
 
-	projects, _ := client.GetProjects(1)
+	projects, _ := client.GetProjects(1, context.Background())
 
 	assert.Len(t, projects, 4)
 	assert.Equal(t, "project-1", projects[0].Name)
@@ -40,7 +42,7 @@ func TestGetProjectsWith2Pages(t *testing.T) {
 func TestGetProjectsWithErrorEmptySlice(t *testing.T) {
 	client := NewClient(mock.NewGitlabClientMock(0, fmt.Errorf("ERROR")))
 
-	projects, _ := client.GetProjects(1)
+	projects, _ := client.GetProjects(1, context.Background())
 
 	assert.Len(t, projects, 0)
 }
