@@ -1,4 +1,4 @@
-const target = 'https://gitlab-ci-dashboard.larscom.nl'
+const target = 'http://localhost:8080'
 
 module.exports = {
   '/api/**': {
