@@ -1,7 +1,7 @@
 import { LatestPipelineStore } from '$groups/group-tabs/feature-tabs/latest-pipelines/store/latest-pipeline.store'
 import { GroupStore } from '$groups/store/group.store'
 import { filterNotNull } from '$groups/util/filter'
-import { CommonModule } from '@angular/common'
+
 import { Component } from '@angular/core'
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop'
 import { FormControl, ReactiveFormsModule } from '@angular/forms'
@@ -14,7 +14,7 @@ import { debounceTime, switchMap, withLatestFrom } from 'rxjs'
 @Component({
   selector: 'gcd-latest-branch-filter',
   standalone: true,
-  imports: [CommonModule, NzIconModule, NzInputModule, NzButtonModule, NzToolTipModule, ReactiveFormsModule],
+  imports: [NzIconModule, NzInputModule, NzButtonModule, NzToolTipModule, ReactiveFormsModule],
   templateUrl: './latest-branch-filter.component.html',
   styleUrls: ['./latest-branch-filter.component.scss']
 })
@@ -23,7 +23,10 @@ export class LatestBranchFilterComponent {
 
   searchControl = new FormControl('')
 
-  constructor(private latestPipelineStore: LatestPipelineStore, private groupStore: GroupStore) {
+  constructor(
+    private latestPipelineStore: LatestPipelineStore,
+    private groupStore: GroupStore
+  ) {
     this.searchControl.valueChanges
       .pipe(takeUntilDestroyed(), debounceTime(100), withLatestFrom(this.selectedGroupId$))
       .subscribe(([value, groupId]) => this.latestPipelineStore.setBranchFilter(groupId, String(value)))
