@@ -1,5 +1,5 @@
 import { UIStore } from '$store/ui.store'
-import { Injectable } from '@angular/core'
+import { Injectable, inject } from '@angular/core'
 import { createEffect, ofType } from '@ngneat/effects'
 import { of, switchMap, tap, zip } from 'rxjs'
 import { LatestPipelineService } from '../service/latest-pipeline.service'
@@ -8,6 +8,10 @@ import { LatestPipelineStore } from './latest-pipeline.store'
 
 @Injectable({ providedIn: 'root' })
 export class LatestPipelineEffects {
+  private latestPipelineStore = inject(LatestPipelineStore)
+  private service = inject(LatestPipelineService)
+  private uiStore = inject(UIStore)
+
   fetchProjectsWithLatestPipeline = createEffect((actions) => {
     return actions.pipe(
       ofType(fetchProjectsWithLatestPipeline),
@@ -31,10 +35,4 @@ export class LatestPipelineEffects {
       tap(([projectId]) => this.uiStore.setAutoRefreshLoading(projectId, false))
     )
   })
-
-  constructor(
-    private latestPipelineStore: LatestPipelineStore,
-    private service: LatestPipelineService,
-    private uiStore: UIStore
-  ) {}
 }
