@@ -1,5 +1,5 @@
 import { UIStore } from '$store/ui.store'
-import { Injectable } from '@angular/core'
+import { Injectable, inject } from '@angular/core'
 import { createEffect, ofType } from '@ngneat/effects'
 import { of, switchMap, tap, zip } from 'rxjs'
 import { PipelineService } from '../service/pipeline.service'
@@ -8,6 +8,10 @@ import { PipelineStore } from './pipeline.store'
 
 @Injectable({ providedIn: 'root' })
 export class PipelineEffects {
+  private pipelineStore = inject(PipelineStore)
+  private service = inject(PipelineService)
+  private uiStore = inject(UIStore)
+
   fetchProjectsWithPipeline = createEffect((actions) => {
     return actions.pipe(
       ofType(fetchProjectsWithPipeline),
@@ -19,6 +23,4 @@ export class PipelineEffects {
       tap(([groupId]) => this.uiStore.setAutoRefreshLoading(groupId, false))
     )
   })
-
-  constructor(private pipelineStore: PipelineStore, private service: PipelineService, private uiStore: UIStore) {}
 }
