@@ -1,4 +1,4 @@
-import { Status } from '$groups/model/pipeline'
+import { Status } from '$groups/model/status'
 import { statusToColor } from '$groups/util/status-color'
 import { Pipe, PipeTransform } from '@angular/core'
 
