@@ -1,4 +1,4 @@
-import { Status } from '$groups/model/pipeline'
+import { Status } from '$groups/model/status'
 import { GroupStore } from '$groups/store/group.store'
 import { filterNotNull } from '$groups/util/filter'
 import { UIStore } from '$store/ui.store'
