@@ -1,5 +1,5 @@
 import { UIStore } from '$store/ui.store'
-import { Injectable } from '@angular/core'
+import { Injectable, inject } from '@angular/core'
 import { createEffect, ofType } from '@ngneat/effects'
 import { of, switchMap, tap, zip } from 'rxjs'
 import { ScheduleService } from '../service/schedule.service'
@@ -8,6 +8,10 @@ import { ScheduleStore } from './schedule.store'
 
 @Injectable({ providedIn: 'root' })
 export class ScheduleEffects {
+  private scheduleService = inject(ScheduleService)
+  private scheduleStore = inject(ScheduleStore)
+  private uiStore = inject(UIStore)
+
   fetchSchedules = createEffect((actions) => {
     return actions.pipe(
       ofType(fetchSchedules),
@@ -17,10 +21,4 @@ export class ScheduleEffects {
       tap(([groupId]) => this.uiStore.setAutoRefreshLoading(groupId, false))
     )
   })
-
-  constructor(
-    private scheduleService: ScheduleService,
-    private scheduleStore: ScheduleStore,
-    private uiStore: UIStore
-  ) {}
 }
