@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core'
+import { Injectable, inject } from '@angular/core'
 import { createEffect, ofType } from '@ngneat/effects'
 import { switchMap, tap } from 'rxjs'
 import { GroupService } from '../service/group.service'
@@ -7,6 +7,9 @@ import { GroupStore } from './group.store'
 
 @Injectable({ providedIn: 'root' })
 export class GroupEffects {
+  private store = inject(GroupStore)
+  private service = inject(GroupService)
+
   fetchGroups = createEffect((actions) => {
     return actions.pipe(
       ofType(fetchGroups),
@@ -14,6 +17,4 @@ export class GroupEffects {
       tap((groups) => this.store.setGroups(groups))
     )
   })
-
-  constructor(private store: GroupStore, private service: GroupService) {}
 }
